Animations are now working in demo + animation data display #32 #40 #41

+++ tests/tests_animation.cpp
@@ -12,26 +12,26 @@
 TEST_CASE( "Create simple animation with 3 frames. Expect right values after update", "[animation]" )
 {
     tson::Animation a {{{100, 1}, {200, 2}, {100, 3}}};
-    a.update(99);
+    a.update(99.f);
     REQUIRE(a.getCurrentFrameNumber() == 0);
     REQUIRE(a.getCurrentFrame()->getDuration() == 100);
     REQUIRE(a.getCurrentFrame()->getTileId() == 1);
-    REQUIRE(a.getTimeDelta() == 99);
+    REQUIRE((int)a.getTimeDelta() == 99);
 
-    a.update(6);
+    a.update(6.f);
     REQUIRE(a.getCurrentFrameNumber() == 1);
     REQUIRE(a.getCurrentFrame()->getDuration() == 200);
     REQUIRE(a.getCurrentFrame()->getTileId() == 2);
-    REQUIRE(a.getTimeDelta() == 5);
+    REQUIRE((int)a.getTimeDelta() == 5);
 
-    a.update(195);
+    a.update(195.f);
     REQUIRE(a.getCurrentFrameNumber() == 2);
     REQUIRE(a.getCurrentFrame()->getDuration() == 100);
     REQUIRE(a.getCurrentFrame()->getTileId() == 3);
-    REQUIRE(a.getTimeDelta() == 0);
+    REQUIRE((int)a.getTimeDelta() == 0);
 
-    a.update(100);
+    a.update(100.f);
     REQUIRE(a.getCurrentFrameNumber() == 0);
     REQUIRE(a.getCurrentFrame()->getTileId() == 1);
-    REQUIRE(a.getTimeDelta() == 0);
+    REQUIRE((int)a.getTimeDelta() == 0);
 }
