+++ tileson.hpp
@@ -5510,18 +5510,18 @@ namespace tson
 			inline Animation() = default;
 			inline Animation(const std::vector<tson::Frame> &frames) : m_frames {frames} {};
 
-			inline void update(int timeDeltaMs);
+			inline void update(float timeDeltaMs);
 			inline void reset();
 
 			inline void setFrames(const std::vector<tson::Frame> &frames);
 			inline void setCurrentFrame(uint32_t currentFrame);
-			inline void setTimeDelta(int timeDelta);
+			inline void setTimeDelta(float timeDelta);
 
 			inline const std::vector<tson::Frame> &getFrames() const;
 			inline const tson::Frame *getCurrentFrame() const;
 			inline uint32_t getCurrentFrameNumber() const;
 			inline uint32_t getCurrentTileId() const;
-			inline int getTimeDelta() const;
+			inline float getTimeDelta() const;
 
 			inline bool any() const;
 			inline size_t size() const;
@@ -5530,7 +5530,7 @@ namespace tson
 			inline int nextFrame();
 			std::vector<tson::Frame> m_frames;
 			uint32_t m_currentFrame {0};
-			int m_timeDelta {0};
+			float m_timeDelta {0};
 	};
 
 	const std::vector<tson::Frame> &Animation::getFrames() const
@@ -5544,7 +5544,7 @@ namespace tson
 	void Animation::reset()
 	{
 		m_currentFrame = 0;
-		m_timeDelta = 0;
+		m_timeDelta = 0.f;
 	}
 
 	/*!
@@ -5565,7 +5565,7 @@ namespace tson
 	 * Update animation based on the fra
 	 * @param timedeltaMs Time in milliseconds
 	 */
-	void Animation::update(int timeDeltaMs)
+	void Animation::update(float timeDeltaMs)
 	{
 		const tson::Frame *frame = getCurrentFrame();
 		if(frame != nullptr)
@@ -5573,7 +5573,7 @@ namespace tson
 			m_timeDelta += timeDeltaMs;
 			if(m_timeDelta >= frame->getDuration())
 			{
-				m_timeDelta = m_timeDelta % frame->getDuration();
+				m_timeDelta = (int32_t)m_timeDelta % frame->getDuration();
 				m_currentFrame = nextFrame();
 			}
 		}
@@ -5584,7 +5584,7 @@ namespace tson
 		return (m_currentFrame+1 >= m_frames.size()) ? 0 : m_currentFrame + 1;
 	}
 
-	int Animation::getTimeDelta() const
+	float Animation::getTimeDelta() const
 	{
 		return m_timeDelta;
 	}
@@ -5609,7 +5609,7 @@ namespace tson
 		m_currentFrame = currentFrame;
 	}
 
-	void Animation::setTimeDelta(int timeDelta)
+	void Animation::setTimeDelta(float timeDelta)
 	{
 		m_timeDelta = timeDelta;
 	}
